mv create_params_dict back to Parameters.jl
