+++ src/Models/Models.jl
@@ -1,7 +1,6 @@
 module Models
 
 include("Parameters.jl")
-include("Utils.jl")
 include("DARWIN/DARWIN.jl")
 include("NiPiZD/NiPiZD.jl")
 
@@ -9,7 +8,7 @@ using .Parameters
 using .DARWIN
 using .NiPiZD
 
-export compute_allometric_parameters
+export compute_allometric_parameters, create_params_dict
 export define_tracer_functions
 
 end # module
